+++ nomos-cli/src/da/network/backend.rs
@@ -28,7 +28,7 @@ use tokio_stream::wrappers::{BroadcastStream, UnboundedReceiverStream};
 use super::swarm::ExecutorSwarm;
 
 type BlobId = [u8; 32];
-type ColumnIdx = u32;
+type ColumnIdx = u16;
 
 const BROADCAST_CHANNEL_SIZE: usize = 128;
